feat(18228): view/pricing cleanup
