--- src/views/Pricing/atoms/currentSubscription.ts
@@ -1,3 +1,0 @@
-import { createBooleanAtom } from '~utils/atoms/createPrimitives';
-
-export const isSubscriptionLoadedAtom = createBooleanAtom(false, 'isSubscriptionLoaded');
