--- src/views/Pricing/mock.ts
@@ -1,8 +1,0 @@
-// TODO: remove when API is ready
-import type { CurrentSubscription } from '~features/subscriptions/types';
-
-export const CURRENT_SUBSCRIPTION_INFO: CurrentSubscription = {
-  id: 'kontur_atlas_edu',
-  billingPlanId: 'test',
-  billingSubscriptionId: 'testtest',
-};
